Move `NposSolution` to frame
