+++ frame/election-provider-support/solution-type/Cargo.toml
@@ -26,4 +26,5 @@ scale-info = "2.0.1"
 sp-arithmetic = { version = "5.0.0", path = "../../../primitives/arithmetic" }
 # used by generate_solution_type:
 sp-npos-elections = { version = "4.0.0-dev", path = "../../../primitives/npos-elections" }
+frame-election-provider-support = { version = "4.0.0-dev", path = ".." }
 trybuild = "1.0.53"
