+++ frame/election-provider-support/solution-type/fuzzer/Cargo.toml
@@ -20,6 +20,7 @@ rand = { version = "0.8", features = ["std", "small_rng"] }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
 scale-info = { version = "2.0.1", default-features = false, features = ["derive"] }
 frame-election-provider-solution-type = { version = "4.0.0-dev", path = ".." }
+frame-election-provider-support = { version = "4.0.0-dev", path = "../.." }
 sp-arithmetic = { version = "5.0.0", path = "../../../../primitives/arithmetic" }
 sp-runtime = { version = "6.0.0", path = "../../../../primitives/runtime" }
 # used by generate_solution_type:
