+++ primitives/npos-elections/src/assignments.rs
@@ -200,10 +200,3 @@ impl<VoterIndex, TargetIndex, P: PerThing> IndexAssignment<VoterIndex, TargetInd
 		})
 	}
 }
-
-/// A type alias for [`IndexAssignment`] made from [`crate::NposSolution`].
-pub type IndexAssignmentOf<C> = IndexAssignment<
-	<C as crate::NposSolution>::VoterIndex,
-	<C as crate::NposSolution>::TargetIndex,
-	<C as crate::NposSolution>::Accuracy,
->;
